Update settings.yml
Batch size of 2k is causing timeouts on collections with large amounts of metadata fields. For example, the International Human Rights Law collection has 34 fields and 2339 items, but times out on export with a batch size of 2k. Reducing to 1k on the deployed instance seems to have fixed, so changing here for now until DEC-1178 is completed.