Few fixes for the adhoc import task
Something weird happens with openuri#open nowadays when it parses the filename that seems to kill the honeypot job. Changed to URI.parse and it preserves the original file name, which seems to work better.

Changed to use Metadata::Setter for setting the item name based on file name. (see https://github.com/ndlib/honeycomb/pull/326)