More containerization
Application changes

- Changed to allow configuring logging via env
- Changed production log level to allow override through env, with :warn
as the default. Previous was :info, which was way to verbose for
production. Also removed the colorizing in the logs in production since
we likely won't be reading these logs via the shell.
- Changed the secrets around a bit to mostly read from env as the example.
Added read from env for the Google secrets and added a mapping for these
in the compose. There's still something not quite right with the post
to export_google_sheet_callback. The gem may need to be updated, which
may require application changes.
- Changed the solr config to support reading env. The previous method of
replacing config values at rails start was not conducive to running rake
tasks that also read solr config from env.
- Changed the devise initializer to skip configuration when we are
precompiling assets. As far as I can tell there's no way to determine this
state from within a config natively in rails, so when precompiling we need
to now add ASSET_PRECOMPILE=true to the env, otherwise we'll get a missing
secrets exception.
- Change AWS security chain handling for creating s3 signed url
- Fixed a client header issue when exporting to Google. Appears to have
been necessary due to a change to the Google APIs that had been made
since we last deployed Honeycomb.
- Changed the sneakers config to read in the necessary secrets from
env and to allow logging to stdout when in a container env.
- Reordered the Raven handling to happen after logging in some cases.
Was having some problems with the Raven call failing, which would prevent
getting info about the problem from logs.

Docker changes

- Added a static SECRET_KEY_BASE to the compose production file so that
the service would fully start in production. This value will different
when deployed. It's only here to run the containers in production mode
locally to test
- Added exit handling to the wait-for-it calls in the rails entry. The
service was failing to connect to the database, but stabilizing anyway
because it was not exiting.
- Due to a [change in Fargate 1.4](https://github.com/aws/containers-roadmap/issues/863),
mounting the precompiled asset volume had to be changed to use a VOLUME
instruction inside of Dockerfile.rails and the nginx container changed
to use volumes_from. Unfortunately, the [volumes_from directive is not available in compose 3](https://docs.docker.com/compose/compose-file/compose-versioning/#version-3).
In order to match how this will be done out in ECS, I changed the
compose version to 2.
- Added a volume to share static images with nginx container.
- This also changes the rails container to share the entire /public dir
as a volume and sym link the images to /mnt/honeycomb. This is so that
we can mount efs to /mnt/honeycomb and still share /public as a volume
to nginx. Nginx will now try to serve the request from /public and fall
back to the upstream rails app if it doesn't find a static file.
- Added an "alb" endpoint in prod compose to terminate ssl when running
locally as production. This is for the okta login/callback to work when
run in a compose env