Add healthchecks
Adds a healthcheck route that explicitly checks for migrations. We will
use this for ECS healthchecks to prevent a new task from serving requests
until migrations are run. This will allow us to get a bit closer to the
way capistrano handles the deploy, ie deploy the new version, run migrate,
then make the new deployment live. This will help us reduce the amount
of time between running migrations and cutting over to the new release.