Fix public images
This fixes an issue with uploading images when deployed to ECS. Rails
expects public/system to exist, but is not kept in git, so image uploads
were failing to get moved by paperclip to the correct dir. Changed the
link of /project_root/public/system/images/honeycomb -> /mnt/honeycomb
to instead be a link of /project_root/public/system -> /mnt/system. The
issue was likely obfuscated locally because I had files in public/system
from a local execution of rails, so the docker build was copying these
in. Changed the dockerignore to ignore this directory so that the
behavior more closely match what the docker context will look like in a
fresh checkout.

Additionally moved away from running rails as root and added a rails user
(uid 999) that will run the processes in the container.