Fix editors in Docker env
The previous deploy strategy expected a config/hesburgh_api.yml to exist
on the host. This file configures how to connect to the Hesburgh API for
searching for users. This PR is to add this file to the Docker build
and instructs rails to read the url and token values from env. This should
fix an issue with searching for users when adding admins to the site, or
when adding editors to a collection.